backend/requirements.txt
Django==5.0.6
djangorestframework==3.16.1
pandas_market_calendars==5.1.1
python-dateutil==2.9.0.post0
requests==2.32.4
requests-futures==1.0.2